packages/backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Task Queue
celery==5.4.0
redis==5.1.1

# Video Processing
opencv-python==4.10.0.84
ffmpeg-python==0.2.0

# Image Processing
imagehash==4.3.1
Pillow==10.3.0

# AI/ML
mediapipe==0.10.14

# Data Validation
pydantic==2.8.0
pydantic-settings==2.4.0

# File Operations
aiofiles==24.1.0
python-magic==0.4.27

# Security
python-jose==3.3.0

# Development
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
anyio==4.4.0
black==24.8.0
flake8==7.1.1
httpx==0.27.0
